+++ hookee/plugins/request_body.py
@@ -2,7 +2,7 @@ from hookee import pluginmanager
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.1.0"
+__version__ = "1.2.2"
 
 plugin_type = pluginmanager.REQUEST_PLUGIN
 print_util = None
@@ -16,12 +16,12 @@ def setup(hookee_manager):
 
 def run(request):
     if request.is_json:
-        print_util.print_basic("Body Type: JSON", fg="magenta")
-        print_util.print_dict("Body", dict(request.json), fg="magenta")
+        print_util.print_basic("Body Type: JSON", color=print_util.request_color)
+        print_util.print_dict("Body", dict(request.json), color=print_util.request_color)
     elif request.form and not request.data:
         print_util.print_basic("Body Type: FORM")
-        print_util.print_dict("Body", dict(request.form), fg="magenta")
+        print_util.print_dict("Body", dict(request.form), color=print_util.request_color)
     elif request.data:
-        print_util.print_basic("Body: {}".format(request.data.decode("utf-8")), fg="magenta")
+        print_util.print_basic("Body: {}".format(request.data.decode("utf-8")), color=print_util.request_color)
 
     return request