+++ hookee/plugins/request_headers.py
@@ -2,7 +2,7 @@ from hookee import pluginmanager
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.1.0"
+__version__ = "1.2.2"
 
 plugin_type = pluginmanager.REQUEST_PLUGIN
 print_util = None
@@ -16,8 +16,9 @@ def setup(hookee_manager):
 
 def run(request):
     if request.headers and "X-Forwarded-For" in request.headers:
-        print_util.print_basic("Client IP: {}".format(request.headers.get("X-Forwarded-For")), fg="magenta")
+        print_util.print_basic("Client IP: {}".format(request.headers.get("X-Forwarded-For")),
+                               color=print_util.request_color)
     if request.headers:
-        print_util.print_dict("Headers", dict(request.headers), fg="magenta")
+        print_util.print_dict("Headers", dict(request.headers), color=print_util.request_color)
 
     return request