+++ hookee/plugins/request_query_params.py
@@ -2,7 +2,7 @@ from hookee import pluginmanager
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.1.0"
+__version__ = "1.2.2"
 
 plugin_type = pluginmanager.REQUEST_PLUGIN
 print_util = None
@@ -16,6 +16,6 @@ def setup(hookee_manager):
 
 def run(request):
     if request.args:
-        print_util.print_dict("Query Params", dict(request.args), fg="magenta")
+        print_util.print_dict("Query Params", dict(request.args), color=print_util.request_color)
 
     return request